fix merge conflict

+++ scenarios/taskrunner/taskrunner.go
@@ -21,13 +21,13 @@ import (
 )
 
 type ScenarioOptions struct {
-	TotalCount  uint64 `yaml:"total_count"`
-	Throughput  uint64 `yaml:"throughput"`
-	MaxPending  uint64 `yaml:"max_pending"`
-	MaxWallets  uint64 `yaml:"max_wallets"`
-	Rebroadcast uint64 `yaml:"rebroadcast"`
-	BaseFee     uint64 `yaml:"base_fee"`
-	TipFee      uint64 `yaml:"tip_fee"`
+	TotalCount  uint64  `yaml:"total_count"`
+	Throughput  uint64  `yaml:"throughput"`
+	MaxPending  uint64  `yaml:"max_pending"`
+	MaxWallets  uint64  `yaml:"max_wallets"`
+	Rebroadcast uint64  `yaml:"rebroadcast"`
+	BaseFee     float64 `yaml:"base_fee"`
+	TipFee      float64 `yaml:"tip_fee"`
 
 	TasksConfig string `yaml:"tasks_config"` // Inline YAML/JSON task configuration
 	TasksFile   string `yaml:"tasks_file"`   // Path to task configuration file or URL
@@ -85,8 +85,8 @@ func (s *Scenario) Flags(flags *pflag.FlagSet) error {
 	flags.Uint64Var(&s.options.MaxPending, "max-pending", ScenarioDefaultOptions.MaxPending, "Maximum number of pending transactions")
 	flags.Uint64Var(&s.options.MaxWallets, "max-wallets", ScenarioDefaultOptions.MaxWallets, "Maximum number of child wallets to use")
 	flags.Uint64Var(&s.options.Rebroadcast, "rebroadcast", ScenarioDefaultOptions.Rebroadcast, "Enable reliable rebroadcast with unlimited retries and exponential backoff")
-	flags.Uint64Var(&s.options.BaseFee, "basefee", ScenarioDefaultOptions.BaseFee, "Max fee per gas to use in transactions (in gwei)")
-	flags.Uint64Var(&s.options.TipFee, "tipfee", ScenarioDefaultOptions.TipFee, "Max tip per gas to use in transactions (in gwei)")
+	flags.Float64Var(&s.options.BaseFee, "basefee", ScenarioDefaultOptions.BaseFee, "Max fee per gas to use in transactions (in gwei)")
+	flags.Float64Var(&s.options.TipFee, "tipfee", ScenarioDefaultOptions.TipFee, "Max tip per gas to use in transactions (in gwei)")
 
 	flags.StringVar(&s.options.TasksConfig, "tasks", ScenarioDefaultOptions.TasksConfig, "Inline task configuration (YAML/JSON string)")
 	flags.StringVar(&s.options.TasksFile, "tasks-file", ScenarioDefaultOptions.TasksFile, "Path to task configuration file or URL (http/https)")
